#Refactored code, Added unit tests, Added Mapstruct

+++ src/main/java/com/example/adverts/controller/AdController.java
@@ -1,13 +1,11 @@
 package com.example.adverts.controller;
-
-import com.example.adverts.domain.Ad;
-import com.example.adverts.domain.Category;
 import com.example.adverts.dto.AdDto;
 import com.example.adverts.service.AdService;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -19,41 +17,61 @@ public class AdController {
     @Autowired
     private AdService adService;
 
-    @RequestMapping(value = "/create", method = RequestMethod.POST)
-    public ResponseEntity<Ad> createAd(@RequestBody AdDto adDto) {
-        Ad ad = adService.createAd(adDto.getText(), adDto.getPrice(), adDto.getExpiryDate(),
-                adDto.getCategory());
-        return ResponseEntity.status(HttpStatus.CREATED).body(ad);
+    @PostMapping("/")
+    public ResponseEntity<AdDto> createAd(@RequestBody AdDto adDto, BindingResult bindingResult) {
+
+        if(bindingResult.hasErrors()){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        AdDto createdAd = adService.createAd(adDto);
+        if(createdAd == null){
+      return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        return new ResponseEntity<>(createdAd, HttpStatus.CREATED);
     }
 
-    @RequestMapping(value = "/update/{id}", method = RequestMethod.PUT)
-    public ResponseEntity<Ad> updateAd(@RequestBody AdDto adDto, @PathVariable("id") Long id) {
-        Ad ad = adService.updateAd(id, adDto.getText(), adDto.getPrice(), adDto.getCategory(), adDto.getExpiryDate());
-        return ResponseEntity.ok(ad);
+    @PutMapping("/{id}")
+    public ResponseEntity<AdDto> updateAd(@PathVariable("id") Long id, @RequestBody AdDto adDto, BindingResult bindingResult ) {
+        if(bindingResult.hasErrors()){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        else if (!adService.findAdById(id).isPresent()) {
+          return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        AdDto updatedAd = adService.updateAd(adDto);
+            return new ResponseEntity<>(updatedAd, HttpStatus.OK);
     }
 
-    @RequestMapping(value = "/get/{id}", method = RequestMethod.GET)
-    public ResponseEntity<Ad> getAdById(@PathVariable("id") Long id) {
-        Ad ad = adService.getAdById(id);
-        return ResponseEntity.ok(ad);
+    @GetMapping("/{id}")
+    public ResponseEntity<AdDto> getAdById(@PathVariable("id") Long id) {
+            if(id == null){
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            }
+            if(adService.findAdById(id).isEmpty()){
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
+        return new ResponseEntity<>(adService.findAdById(id).get(), HttpStatus.OK);
     }
 
-    @RequestMapping(value = "/remove/{id}", method = RequestMethod.DELETE)
-    public void deleteAdById(@PathVariable Long id) {
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteAdById(@PathVariable Long id) {
+        if(adService.findAdById(id).isEmpty()){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         adService.deleteAdById(id);
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
-    @RequestMapping(value = "/get-all", method = RequestMethod.GET)
-    public ResponseEntity<List<Ad>> getAllAds() {
-        List<Ad> ads = adService.getAllAdds();
-        return ResponseEntity.ok(ads);
+    @GetMapping("/")
+    public ResponseEntity<List<AdDto>> getAllAds() {
+        List<AdDto> ads = adService.getAllAdds();
+        return new ResponseEntity<>(ads, HttpStatus.OK);
     }
 
-    @RequestMapping(value = "/filter/{category}", method = RequestMethod.GET)
-    public ResponseEntity<List<Ad>> filterAdsByCategory(@PathVariable String category) {
+/*    @GetMapping("/filter/{category}")
+    public ResponseEntity<List<AdDto>> filterAdsByCategory(@PathVariable String category) {
         List<Ad> filteredAds = adService.filterAdsByCategory(Category.valueOf(category.toUpperCase()));
         return ResponseEntity.ok(filteredAds);
-    }
-
+    }*/
 
 }
